+++ flake.nix
@@ -69,6 +69,7 @@
           flake = fixtureFlake;
           title = "Example fleet";
           domains.ts = "ts.example";
+          theme = "light";
         };
 
         site =
@@ -80,8 +81,9 @@
               cp -r ${./site} book
               chmod -R u+w book
               cp ${./SYNTAX.md} book/src/syntax.md
-              cp ${./assets}/topology.svg ${./assets}/modules.svg book/src/
-              cp ${./tests/reference}/wiki/src/closures.svg book/src/
+              cp ${./assets}/topology-light.svg book/src/topology.svg
+              cp ${./assets}/modules-light.svg book/src/modules.svg
+              cp ${./assets}/closures-light.svg book/src/closures.svg
               mdbook build book --dest-dir $out
               cp -r --no-preserve=mode ${demo-docs}/wiki/book $out/demo
             '';